BUG/MEDIUM: thread/http: Add missing locks in set-map and add-acl HTTP rules
Locks are missing in the rules "http-request set-map" and "http-response
add-acl" when an acl or map update is performed. Pattern elements must be
locked.

This patch must be backported to 1.9 and 1.8. For the 1.8, the HTX part must be
ignored.

(cherry picked from commit e84289e5854aa3b00cd19032387f435ca6748491)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>