BUG/MEDIUM: connections: Make sure we remove CO_FL_SESS_IDLE on disown.
When for some reason the session is not the owner of the connection anymore,
make sure we remove CO_FL_SESS_IDLE, even if we're about to call
conn->mux->destroy(), as the destroy may not destroy the connection
immediately if it's still in use.
This should be backported to 1.9.
u

(cherry picked from commit a48237fd0707a78a353613a45130f4b433b19b0f)
Signed-off-by: Willy Tarreau <w@1wt.eu>