BUG/MEDIUM: streams: Don't add CF_WRITE_ERROR if early data were rejected.
In sess_update_st_con_tcp(), if we have an error on the stream_interface
because we tried to send early_data but failed, don't flag the request
channel as CF_WRITE_ERROR, or we will never reach the analyser that sends
back the 425 response.

This should be backported to 1.9.

(cherry picked from commit f4bda993dde4f7414a59d48cb3b009a29456d4ba)
Signed-off-by: Willy Tarreau <w@1wt.eu>