MEDIUM: mux-h2: discard contents that are to be sent after a shutdown
In h2_snd_buf() we discard any possible buffer contents requested to be
sent after a close or an error. But in practice we can extend this to
any case where the stream is locally half-closed since it means we will
never be able to send these data anymore.

For now it must not change anything, but it will be used by subsequent
patches to discard lone a HTX EOM block arriving after the trailers
block.

(cherry picked from commit 2b778484189735d22c45034b22de4665a745c49d)
[wt: needed by next patch]
Signed-off-by: Willy Tarreau <w@1wt.eu>