BUG/MINOR: mux-h1: Fix the parsing of trailers
Unlike other H1 parsing functions, the 3rd parameter of the function
h1_measure_trailers() is the maximum number of bytes to read. For others
functions, it is the relative offset where to stop the parsing.

This patch must be backported to 1.9.

(cherry picked from commit 39593e6ae3442babab462c0b94abf6cc8c755146)
Signed-off-by: Willy Tarreau <w@1wt.eu>