BUG/MEDIUM: h2/htx: never leave a trailers block alone with no EOM block
If when receiving an H2 response we fail to add an EOM block after too
large a trailers block, we must not leave the trailers block alone as it
violates the internal assumptions by not being followed by an EOM, even
when an error is reported. We must then make sure the error will safely
be reported to upper layers and that no attempt will be made to forward
partial blocks.

This must be backported to 1.9.

(cherry picked from commit 2135f91d18ae79c5360d32dd8bf10f42aa6ff8b4)
Signed-off-by: Willy Tarreau <w@1wt.eu>