BUG/MINOR: mux-h2: Count EOM in bytes sent when a HEADERS frame is formatted
In HTX, when a HEADERS frame is formatted before sending it to the client or the
server, If an EOM is found because there is no body, we must count it in the
number bytes sent.

This patch must be backported to 1.9.

(cherry picked from commit 316934d3c905907e8dc8aadec3d1be237177bf7f)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>