BUG/MEDIUM: proto-htx: Not forward too much data when 1xx reponses are handled
When an 1xx reponse is processed, we forward it immediatly. But another message
may already be in the channel's buffer, waiting to be processed. This may be
another 1xx reponse or the final one. So instead of forwarding everything, we
must take care to only forward the processed 1xx response.

This patch must be backported to 1.9.

(cherry picked from commit f90c24d14c6ea541717d92d140f2b4c3f3795117)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>