BUG/MEDIUM: connection: Use the session to get the origin address if needed.
In conn_si_send_proxy(), if we don't have a conn_stream yet, because the mux
won't be created until the SSL handshake is done, retrieve the opposite's
connection from the session. At this point, we know the session associated
with the connection is the one that initiated it, and we can thus just use
the session's origin.

This should be backported to 1.9.

(cherry picked from commit 661167d136d5bf4ac40ac4fae450d66dd69a64df)
Signed-off-by: Willy Tarreau <w@1wt.eu>